java refactor v1

+++ Pizzeria/src/main/java/com/example/Pizzeria/repository/CrustRepository.java
@@ -7,5 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CrustRepository extends JpaRepository<Crust,Integer> {
-    List<Crust> findByCrustOrPrice(String crust,float price);
+
+    List<Crust> findByCrustOrPrice(String crust,Double price);
+
 }
